Fix incorrect Print.write() return value.
The write method must return the number of byte successfully written
(1 in this case).

This fixes the library with the latest AVR core that stops writing
on error:
arduino/Arduino@25d81c9

Without this patch the LCD shows only the first charater of a string.

(Same fix as this one:
https://github.com/marcoschwartz/LiquidCrystal_I2C/commit/143b5a1cb34be3297b6185c58a218779499a2d67)